apps/content_briefs/requirements.txt
Flask
celery
redis
requests
beautifulsoup4
openai==1.8
praw
python-dotenv
serpapi
python-docx 